backend/requirements.txt
fastapi
uvicorn[standard]
feedparser
python-dotenv
huggingface_hub
httpx
google-genai
